Jarvis/requirements.txt
# Core
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0

# Voice & Audio
speechrecognition>=3.10.0
pyttsx3>=2.90
